Change job id values to be numbers generated randomly by default
This is necessary when we are passing the value to the container. With the meaningful string values, dependabot updater attempts to convert to number and gets NaN. The CLI did not pass this value but now that we have an API server that may have more functionality, this is better off as a number.