Add tool usage tracking
The information collected is shown in the logs at the end of each job. Nothing sensitive.

This is to help me understand the usage of the tool/extension/cli through a couple of weeks. There is no other source of this information except for when reviews are left but no-one is bothered to leave those or to even want to give feedback via other channels.
Do people still use this? Let's find out.

Example payload
```json
{
    "trigger": "user",
    "provider": "azure",
    "owner": "https://dev.azure.com/tingle/",
    "package-manager": "terraform",
    "host": {
        "platform": "darwin",
        "release": "25.0.0",
        "arch": "arm64",
        "machine": "d3bbb66be2ad9dfab10af69b450f7e7e814ef7bbf1277a6d0df9e1db44ba4f5c"
    },
    "version": "0.9.0",
    "id": 4210136136,
    "started": "2025-10-03T14:54:25.565Z",
    "duration": 29307,
    "success": true
}
```